requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.11.12
charset-normalizer==3.4.4
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
libclang==18.1.1
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.3
numpy==2.1.2
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==11.3.0
protobuf==6.33.1
Pygments==2.19.2
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.1
